Add files via upload
Goodreads Dataset with 11127 rows and 13 columns. 
BookID: The unique ID for each book/series
Title: The title of the book
Authors: The author of the particular book
Average_raiting: The average rating of the books, as decided by the users
ISBN: The ISBN number, gives detailed information about a book - such as edition and publisher
ISBN 13: The new format for ISBN, implemented in 2007. 13 digits
Language_code: Tells in which language the book is written
Num_pages: The number of pages of the book
Ratings_count: Contains the number of ratings given for the book
Text_reviews_count: Has the count of reviews left by users
Pubication_date: The date when the book was published
Publisher: The publisher of the book
Price: The price in $ of the book
Genre: The genre of the book